+++ services/user/src/modules/auth/authRoutes.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import authController from "../controllers/authController";
+import authController from "./authController";
 
 const router = express.Router();
