+++ services/user/src/server.ts
@@ -2,8 +2,10 @@ import express, { Express } from "express";
 import cors from "cors";
 import dotenv from "dotenv";
 import mongoose from "mongoose";
-import authRoutes from "./routes/authRoutes";
-import userRoutes from "./routes/userRoutes";
+import authRoutes from "./modules/auth/authRoutes";
+import userRoutes from "./modules/user/userRoutes";
+import addressRoutes from "./modules/address/addressRoutes";
+import paymentMethodRoutes from "./modules/paymentMethod/paymentMethodRoutes";
 import { createServerRouter, httpLogger, logger, rabbitMQService } from "@eloritzkovitz/server-essentials";
 
 const app = express();
@@ -19,6 +21,8 @@ app.use(httpLogger);
 // Serve API routes
 app.use("/auth", authRoutes);
 app.use("/users", userRoutes);
+app.use("/users", addressRoutes);
+app.use("/users", paymentMethodRoutes);
 
 // Serve info/health routes
 app.use("/", createServerRouter(() => mongoose.connection.readyState === 1 ? "up" : "down"));
